Set up UART2 interrupt flag to be set when one character is received.
The problem that we had previously was that the commands were not being
read one by one. This was because the ISR was only being triggered when
four characters had been received and, if the command end mark was not
among those characters, the commands was not identified until the next
command was sent.

To fix this, we have changed the 'hostcom_setup' function so that it sets
up the UART2 to interrupt when a single characters arrives (i.e. not to
wait until four are received). In addition, the UART2 receive ISR has been
adapted accordingly.