Implement PWM module with a function to set it up and another one to generate a PWM signal on PWM output #1.
The desired period for the PWM signal is 20 ms (as suggested in [2]), but
from the measurements in the oscilloscope it seems that the actual period
of the generated PWM signal is 2 ms instead. This would suggest that any
of the configuration values for the microcontroller's PWM module is wrong.

In addition, setting a duty cycle of any percentage controls the PWM on
time perfectly, and setting a duty cycle of 100% with an amplitude of 20 V
will cause the motor to rotate at full speed. However, setting a duty
cycle other than 100% will cause the motor to rotate faster some times and
slower other times, probably being faster when the PWM signal is on and
slower when the signal is off (because it's braking). Thus, the speed of
the motor is not even (or homogeneous) over time. This is likely an
indication that the period of the PWM signal should be much shorter.

Besides, the documentation found on how to use the PWM module of the
microcontroller is incomplete and confusing. On the one hand, the official
documentation (datasheets) from Microchip is not complete or does not
provide enough detail on the usage and behavior of the PWM module. On the
other hand, the websites that provide information or code examples ([1],
[2], [3]) don't have enough commentaries on why the code is written like
that and on why those values are chosen, and sometimes even contradict
each other.

[1]: http://www.microchip.com/forums/m338658-print.aspx
[2]: http://www.foxytronics.com/learn/microcontrollers/dspic/dspic30f-motor-control-pwm-module/calculations
[3]: http://batchloaf.wordpress.com/2010/12/06/motor-pwm-example-for-the-dspic30f4011/