Fix to adjust PWM period to 20 ms.
First of all, in the previous code, the 'PTPER' register was not being set
to the value corresponding to a PWM period of 20 ms.

Besides, the calculated value for 'PTPER' was too big to fit into the 15
bits of the 'PTPER' register with a prescale of 1. Thus, a prescale of 16
is now being used.

Finally, the calculation of the value for the 'PDC1' register (duty cycle)
was not being done correctly. Previously, it was being done under the
consideration that the maximum value for the register was 65535, which was
taken from [1]. However it seems that this information has been
misinterpreted and that this is the maximum value that the register can
hold, but not the value that indicates a duty cycle of 100%. Thanks to
measurements with the oscilloscope, we have derived a value of two times
the value of 'PTPER' to indicate a 100% duty cycle. In [1] again, the
value assigned for the 100% duty cycle is two times the value of 'PTPER',
which coincides with our derivation. However, I still don't know why it
has to be two times the value of 'PTPER'; I just know it works.

These changes solve the problem that the PWM period didn't match the
desired period, but the issue with the uneven motor velocity persists.

[1]: http://batchloaf.wordpress.com/2010/12/06/motor-pwm-example-for-the-dspic30f4011/