Change the PWM period from 20 ms to 18 ms to solve a timing issue in the PWM signal.
Previously, when the PWM period was set to 20 ms, both the hardware and
the software PWM modules were outputting a signal with a period a bit
longer than 20 ms, probably due to a mismatching clock frequency.

By lowering the nominal PWM period to 18 ms, both the hardware and
software PWM modules output a PWM signal with a period of approximately
20 ms (the hardware module may output a period slightly less than 20 ms
and the software module may output a period slightly longer than 20 ms,
but the period is now more accurately near 20 ms than it was before).