Fix issue with the frequency of timer 2.
The value to be stored in the period register of timer 2 (PR2) was being
saved to the register PR1 instead. This has been fixed. Additionally, the
macros T2PERIOD and T2PR1VAL have been renamed to T2FREQ and PR2VAL.
Besides, the frequency of timer 2 has been raised from 10 kHz to 100 kHz.