Solve the issues with the 'mcuicom' interpreter.
This commits introduces some changes made to solve the issues with the
microcontroller inter-communication. This includes various different
attempts to fix these issues, which makes the code be a bit messy.

Most of the code is commented out and must be cleaned. However, a specific
piece of code may be useful, though not vital: the code that implements a
timeout for the 'mcuicom'.