Adapt 'mctlcom_get_response' and 'hostcmd_pa' to be used with ASCII commands.
'hostcmd_pa' has been fixed to translate the host command 'PA' into the
corresponding MCUICOM command 'RA', 'RB', etc. depending on the motor
letter.

'mctlcom_get_response' has been modified to return the received response
as-is (as an ASCII string, without converting it to an integer number),
since conversion is not necessary, because the host PC expects the
response as an ASCII string.