Fix lexical parser to not drop the first digit of an integer number.
Previously, the lexical parser was consuming one symbol (the first symbol)
of the input string without saving it if it wasn't a minus sign, which
caused the loss of the first digit of positive integer numbers.

This was affecting, e.g., the CG command, which was always getting a
zero-valued parameter even though the specified parameter was 1.