Implement a test hard home routine.
This routine doesn't really do a hard home. Rather, it checks the limit
switch until it is active and then makes the motor rotate until the limit
switch is found again (i.e. approx. 360 degrees).

This is done to verify that the limit switch circuitry and the QEI
software component work well (mostly that the QEI counts the steps
correctly, which can be done by comparing the steps counted by the
microcontroller with the number of steps calculated from the formulae).