Move initial state out of the ISR and rename states.
The initial state of the automaton has been moved out of the ISR, to the
hard home routine, because enabling the Change Notification interrupts was
seemingly not causing the interrupt to be triggered. Thus, the actions
initial state, which starts the hard home process, had to be performed in
the hard home routine, so thath the motor begins to move and actually
causes the interrupts to be triggered.

In addition, the states have been renamed to use more descriptive (and
gramatically correct) names (i.e. where the name describes the action that
the automaton/microcontroller is performing).