Start implementation of a hard home routine based in the GPMCU.
This hard home routine runs on the GPMCU, which controls the limit
switches. The GPMCU communicates with the MCMCU sending commands to move
the motors. The GPMCU then checks the limit switches and, when they are
activated, it tells the MCMCU to stop the motors. In other words, the
MCMCU doesn't have any knowledge of the limit switches, it just obeys the
commands from the GPMCU.