Implement MCMCU command to reset the PWM parameters to the default values.
The idea is to have a procedure that resets the PWM parameters to the
values stored in the EEPROM, to be used mainly after a hard home, which
modifies the PWM parameters. However, since the EEPROM module is not yet
implemented, this procedure simply resets the PWM level to 100% duty cycle
and the PWM direction to 0.