Decrease frequency of Timer 3 (PID loop interrupt) to 10 KHz.
Timer 3 was running at 100 KHz, like the QEI timer (Timer 2). The MCMCU
was not responding to any commands. It is suspected that this was because
the Timer 3 was running so quickly that it prevented the main loop to
execute, as found out by putting a 'printf' into the main loop and
checking that nothing was bein printed out.

To solve this issue, the frequency of Timer 3 was decreased to 10 kHz.
This frequency (and the frequency of Timer 2) could be reduced even more,
since the maximum frequency of the encoder signalsd is 2 kHz. Thus, it is
suggested that the Timer 2 frequency be reduced to around 10 kHz and the
Timer 3 frequency to around 5 kHz or 2 kHz at most.