Add new motor status register 'motor_commanded_pos' to store the commanded position, in distinction to the motor desired position.
The motor commanded position is the position that the motor has been
commanded to go to, but where it is not going to until an MI or MC command
is issued. This is used to perform the trapezoidal move to that commanded
position.

The desired position is the position that the motor is already in after
the trapezoidal move has finished. This register is used by the PID loop
to know the position where the motor should be in order to compensate
disturbances and thus keep the motor in the desired position.