Fix initialization of the QEI module so that the step count is always 0 on startup.
Since these instructions were missing, the step counter used to be
different from zero, because the first time the QEI ISR was executed, it
detected a change in the encoder state and thus counted one step.

With these instructions for the initialization of the current encoder
state, the ISR doesn't detect any change until the motor actually moves,
thus eliminating the error.