Implement an MCMCU command to check if a given motor has stopped moving.
This command is implemented by checking a certain number of times if there
is any change in the motor step counter register or not. If there is no
change for a certain number of times, it is considered that there has been
no change in the motor position. At that point, the command returns.
Otherwise, it blocks until no change is detected.

This command will serve to support the hard home routine.