Fix hard home routine: wait until motor stops before clearing position register.
If the position register is cleared without waiting for the motor to stop,
the position will be reset to zero while the motor is still moving and
thus the movement to the mid-point will not be executed correctly. In
fact, the motor will move away from the mid-point, thus making the hard
home defective.