Fix hard home routine to finish at the correct motor position.
The problem wasn't that the motor position register was being cleared
before getting to the correct final position, but rather that the motor
desired position wasn't being cleared as well, and this was causing the
PID loop to compensate the position error and thus move the motor past the
correct hard home position.

After this change, the MCMCU command to check if a certain motor has
stopped moving is no longer needed.