Various fixes to prevent sign errors.
The trapezoidal velocity profile generation algorithm no longer checks if
it has surpassed the mid-point of the trajectory by actually calculating
the mid-point, but rather by counting the number of steps given and
checking if the number of steps given is greater than the number of steps
necessary to reach the mid-point from the starting point of the
trajectory. This will help prevent sign errors when the initial and final
positions have different sign.

The fields of the 'motorctl_info_t' structure have been declared as
'integer' instead of 'unsigned integer', because integer addition and
subtraction is performed with several of these fields throughout the code
of the algorithm, and overflows and underflows could occur when dealing
with limit values of the ('unsigned integer') arithmetic, leading to
errors in the comparisons (less than, greater than). By making these
fields 'integer' instead of 'unsigned integer', we make sure that there
are no underflows when subtracting, or overflows when adding, thus
preserving the result of the comparisons.