Fix trapezoidal velocity profile calculation algorithm.
The counters used in the algorithm need to be updated at the same rate
(i.e. timer frequency) than the step counters (QEI) if they are to be
compared with each other. Otherwise, the counters of the algorithm won't
hold a count of the motor steps given.

To solve this, the absolute value of the number of steps that need to be
given to reach the mid-point of the trajectory is used. This enables the
algorithm to work at an arbitrary frequency, different from the QEI timer
frequency.