Change proportional gain of P control to 10.
Currently, only a proportional control is implemented (no integral or
derivative terms). The P gain has been changed from 1 to 10 in order to
try if that makes the trapezoidal movement a bit faster and
non-intermitent, which is how it behaves right now.

However, the downside of this is that the hard home is a lot faster, which
is not very appealing to the eye and might lead to position inaccuracies
when handling the limit switch.

Also, the trapezoidal movement behaves a lot worse: equally intermitent
(as it was before) but a lot more abruptly. After commanding a movement of
1000 steps, it ends at 1200 steps instead, as it used to do before the
changes of this commit.