Change PID gains to P=1, D=1 with no integral term (I=0).
The movement is a lot smoother, but also much slower. Besides, the PWM
(and, thus, the motor) don't stop when it arrives to the destionation
position. Instead, it keeps correcting its position (oscillating), though
very slightly, since it's already arrives at the desired position.

Also, when commanded to move 1000 steps, it ends up at around 1400 steps.
We have been wondering how to solve this, but it has just occurred to us
that this may be due to some steady state error in the PID control, since
the integral term has been zeroed.