Change PID gains to P=1, I=1, D=0.
The hard home is somehow smooth, but it oscillates further from the limit
switch before it finally arrives at the mid-point of the switch (probably
due to the integral term introducing oscillations and overshoot). However,
the trapezoidal movement is still intermitent and abrupt, and the offset
from the commanded position remains (at least it remained in the first
movement we tried, but then no offset was present in a second movement we
commanded).