Fix trapezoidal movement to support negative movements.
Previously, only positive movements were supported. When a negative
movement was attempted, the controller also performed a positive movement,
because the velocity was always being added to the position.

With this fix, the velocity is added to the position when the movement is
positive, and subtracted if the movement is negative.