Fix to correct motor position after finishing the movement.
The code to enable the trapezoidal velocity profile generation for unused
motors has been commented out (only during the tests with motor A). These
were preventing the program to reach the if-block where 'move_finished' is
true and the desired position is set equal to the commanded position, thus
correcting any error in the position (i.e. the offset caused by the
trapezoidal velocity profile generation).

Not only has this code been commented out, but a new if-block has been
added to set the desired position equal to the commanded position after
the movement of motor A has finished, thus not having to wait until all
motores finish. This helps to fix the issue with the position offset when
a motor ends its movement earlier than other motors.

However, this doesn't eliminate the offset. The offset is still present,
but is corrected at the end by the PID control loop. Moreover, the PID
loop corrects the offset very quickly, which causes a very ugly and abrupt
movement backwards at the end of the trapezoidal move. This could probably
be mitigated with a derivative term in the PID control, but the offset
needs to be investigated and eliminated anyway.