Fix 'motorctl' to not wait for the trapezoidal move to finish, so that the MCMCU will be able to respond to any command from the GPMCU.
This enables the MCMCU to respond to the GPMCU when it asks for the motor
status bits, for example.

This active wait is not necessary, because the GPMCU already enforces that
no movement can be performed while the MCMCU is executing a trapezoidal
move. Thus, move commands like PR or PD, which modify the commanded
position register, will have no effect on the movement that is currently
executing. (Before implementing the communication between the GPMCU and
the MCMCU to synchronize the motor status register, these commands would
modify an already executing movement, because they change the commanded
position register.)