Fix deadlock between GPMCU and MCMCU when reading the PID gains of a motor.
When the MCMCU received any of the commands to read one of the gains of a
motor, it was checking for the presence of a second parameter, which is
not provided according to the format of the commands. Thus, the MCMCU was
not sending back any response to the GPMCU.

Since the GPMCU expected a response from the MCMCU in a blocking fashion
and no timeout has been built-in yet, the GPMCU didn't continue the
execution of the program.

In this commit we have removed the requirement of a second parameter to
these commands, thus eliminating the deadlock issue.