Fix to set the PID gains from the GPMCU instead of the MCMCU.
The commands to read the PID gains where always returning zero, even
though the MCMCU sets the PID gains on startup and, at the time being, the
proportional gain of motor A was non-zero.

In this commit, we shift the task of setting up the PID gains on startup
from the MCMCU to the GPMCU, and this solves the issue.