Fix error in the functions that are used to get the PID gains in the MCMCU (same error as in the previous commit).
This was causing the RA, RB and RC commands to return funny values for the
PID gains: e.g. RA was returning zero for motor A (which had proportional
gain 1) and -1 for motors D, E and F (which is the default value of the
switch-block).

Even though in the commit message of commit 64c7a71 was stated that that
commit solved the issue, it actually didn't. Now, THIS commit does solve
the issue.